Add type generators and mapper into separated file
