+++ src/buildProps.js
@@ -1,18 +1,4 @@
-const generateString = () => 'abs';
-const generateBool = () => true;
-const generateNumber = () => 123;
-const generateIntlShape = () => 'tricky';
-const generateInstanceOf = () => 'tricky';
-const generateWidthPropType = () => '100';
-
-const mapper = {
-  string: generateString(),
-  bool: generateBool(),
-  number: generateNumber(),
-  intlShape: generateIntlShape(),
-  instanceOf: generateInstanceOf(),
-  WidthPropType: generateWidthPropType(),
-}
+import { mapper } from './typeValueGenerators';
 
 const buildProps = (props, propTypes = {}, shapeProp, shapeRestListCount) => {
   if (props.length === 0) {
